% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_is_df}
\alias{assert_is_df}
\title{assert_is_df aborts program if object isn't a data.frame-like class}
\usage{
assert_is_df(
  object,
  error_msg =
    "ECB Data API response is not a valid CSV file. Something must have gone very wrong. Most likely this is not your fault."
)
}
\arguments{
\item{object}{some object}

\item{error_msg}{\emph{message to show when program aborts}

\verb{scalar<character>}  //  \emph{default:} \verb{ECB Data API response is not a valid CSV file. Something must have gone very wrong. Most likely this is not your fault.} (\code{optional})

in this package this function is only used for ECB Data API response... therefore the standard message}
}
\value{
nothing, but may abort if object is not a data.frame-like class
}
\description{
assert_is_df aborts program if object isn't a data.frame-like class
}
\examples{
assert_is_df("hfio", "This is not a df")
}
\keyword{internal}
