% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_mturk.R
\docType{data}
\name{bench_mturk}
\alias{bench_mturk}
\title{Benchmark Dataset based on Mturk Samples}
\format{
data.frame
}
\usage{
bench_mturk
}
\description{
Benchmark Dataset based on Mturk Samples
}
\keyword{dataset}
