% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_prob_survey.R
\docType{data}
\name{bench_prob_survey}
\alias{bench_prob_survey}
\title{Benchmark Dataset based on National Probability Surveys}
\format{
data.frame
}
\usage{
bench_prob_survey
}
\description{
Benchmark Dataset based on National Probability Surveys
}
\keyword{dataset}
