% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{constraint}
\alias{constraint}
\title{Adding Constraints}
\usage{
constraint(vars, type, values)
}
\arguments{
\item{vars}{Names of variables that users want to add constraints.}

\item{type}{One of the three types; "between", "less than or equal to", "larger than or equal to".}

\item{values}{Values of the weighted mean of variables specified in `vars`. When `type = "between"`, `values` should be a vector of length 2. Otherwise, `values` should be one numeric value.}
}
\description{
Adding Constraints
}
