% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr.R
\name{exr}
\alias{exr}
\title{Estimating External Robustness}
\usage{
exr(
  outcome,
  treatment,
  covariates,
  data,
  sate_estimate = NULL,
  family = "gaussian",
  threshold = 0,
  cate_method = "grf",
  cate_name = NULL,
  uncertainty = TRUE,
  ci = 0.95,
  boot = FALSE,
  nboot = 100,
  clusters = NULL,
  numCores = 1,
  const_list = NULL,
  lib = c("mean", "glm", "ranger"),
  verbose = FALSE,
  seed = 1234
)
}
\arguments{
\item{outcome}{Name of the outcome variable.}

\item{treatment}{Name of the treatment variable. The treatment variable needs to be a binary variable.}

\item{covariates}{Name of the covariates}

\item{data}{`data` needs to be `data.frame`}

\item{sate_estimate}{A vector of length 2; a point estimate of the SATE and its standard error. Default = NULL. When NULL, the package internally estimates the SATE using a linear regression of the outcome on the treatment and all specified covariates.}

\item{family}{"gaussian" (continuous outcomes) or "binomial" (binary outcomes). Default = `gaussian`.}

\item{threshold}{Numeric. This is the threshold for the T-PATE. Default = 0}

\item{cate_method}{When "grf", we use 'grf'. When "X-learner", we use 'X-learner' based on R package 'SuperLearner'.}

\item{cate_name}{Name of columns in `data` that store CATEs estimated outside the function. Default = NULL, and the CATE is internally estimated by a method specified by `cate_method.`}

\item{uncertainty}{Logical. Whether we take into account uncertainties to estimate external robustness. Default is TRUE.}

\item{ci}{Numeric. Specify the level of the confidence interval. Default = 0.95. Only used when `uncertainty = TRUE`.}

\item{boot}{Logical. When `boot = TRUE`, we use bootstrap to approximate confidence intervals. When `boot = FALSE`, we approximate the standard error using the standard error of the SATE (This is computationally much faster, so researchers can use this as an initial check. But for final results, we recommend `boot = TRUE`).}

\item{nboot}{Numeric. The number of bootstrap. Default = 100. Only used when `uncertainty = TRUE` and `boot = TRUE`.}

\item{clusters}{Vector. Unique identifiers for computing cluster standard errors. Only used when `uncertainty = TRUE` and `boot = TRUE`.}

\item{numCores}{Numeric. Default = 1. Number of cores to use. Parallel computing based on `future` package is used only when `uncertainty = TRUE` and `boot = TRUE`. When `numCores = NULL`, it automatically detects the number of available cores.}

\item{const_list}{List. Constraints to incorporate partial knowledge about population data. See Examples.}

\item{lib}{(optional) The library used for 'SuperLearner' when we choose `cate_method = "X-learner"`. Default = c("mean", "glm", "ranger").}

\item{verbose}{Logical. Whether to see outputs from the underlying optimization package `CVXR`. Default = FALSE.}

\item{seed}{Numeric. `seed` used internally. Default = 1234.}
}
\value{
\code{exr} returns an object of \code{exr} class.
 \itemize{
   \item \code{est}: Estimated external robustness.
   \item \code{est_unc}: Estimates of the pAMCE for all factors in each bootstrap sample.
   \item \code{tau}: Estimated conditional average treatment effects (CATEs) for each unit.
   \item \code{w}: Estimated weights that solves the KL minimization problem where a constraint focuses on a point estimate of the T-PATE.
   \item \code{w_unc}: Estimated weights that solves the KL minimization problem where a constraint focuses on a confidence internal of the T-PATE.
   \item \code{...}: Values for internal use.
 }
}
\description{
Estimating External Robustness
}
\references{
Devaux and Egami. (2022+). Quantifying Robustness to External Validity Bias. Available at \url{https://naokiegami.com/paper/external_robust.pdf}.
}
