% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_cross_rates.R
\name{get_all_cross_rates}
\alias{get_all_cross_rates}
\title{returns the cross-rates (or exchange rates) for a specific day}
\usage{
get_all_cross_rates(
  currencies = NULL,
  date = Sys.Date() - 1,
  max_lookback_days = 7,
  output_method = "long"
)
}
\arguments{
\item{currencies}{\emph{currency or currency list}

\verb{scalar<character>} or \verb{vector<character>} // \emph{default:} \code{NULL} for all (\code{optional})

If you want specific currencies, provide them to the function as a vector, such
as c("USD","THB")}

\item{date}{\emph{the as-of-date}

\verb{scalar<Date>} //  \emph{default:} \code{Sys.Date()-1} (\code{optional})

Data will be displayed for this date. If there is no data for this sopecific date,
the package will go back in history and look for most recent exchange rates. You
can limit this by \code{max_lookback_days}}

\item{max_lookback_days}{\emph{look back if missing for not more than X days}

\verb{scalar<numeric>} //  \emph{default:} \code{7} (\code{optional})

At \code{date} there might be missing data because ECB will only publish data on working days.
The values in \code{EXR::get_all_cross_rates} are taken from the most recent day before. If
data is too old, you might not want to use outdated exchange rates though. Set the maximum
number of calender days you want to go back with \code{max_lookback_days}. If you don't want to use
data of previous days, set this parameter to \code{0} or NULL}

\item{output_method}{\emph{how do you want the output to be formatted}

\verb{scalar<character>} //  \emph{default:} \code{long} (\code{optional})

\itemize{
\item \strong{long}/\strong{deep}: output will be a deep table with columns base_currency, price_currency, value
\item \strong{wide}/\strong{pivot}: output will be a wide (pivot) table with base_currency as rows and price_currency as columns, value as combination
}}
}
\description{
get_all_cross_rates will take a number of currencies as input and return
all exchange rates between them. If the ECB does not publish them, cross-rates
will be calculated based on EUR-based currency pairs
}
\examples{
# get all possible cross-rates / exchange rates as of yesterday
EXR::get_all_cross_rates()

# get all combinations for a set of currencies as of end 2023 and show as pivot table
EXR::get_all_cross_rates(currencies = c("EUR", "USD", "JPY", "GBP", "CNY"), date = as.Date("2023-12-31"), output_method = "pivot")

}
