% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_currencies.R
\name{get_available_currencies}
\alias{get_available_currencies}
\title{returns a \code{tibble} of available currencies and their English names}
\usage{
get_available_currencies(date = Sys.Date() - 1, use_most_recent = TRUE)
}
\arguments{
\item{date}{\emph{the 'as of'-date}

\verb{scalar<Date>} // \emph{default:} \code{Sys.Date()-1} (\code{optional})

The API will be searched for available currencies at this specific date. Therefore it is not
allowed to query data in the future. If you are not sure whether there is any data at all for
this specific date (e.g. weekend), use param \code{use_most_recent} to get a non-empty result
(if \code{date} is not out of minimum date boundary of the database). Yesterday was chosen as
default because during a day, the results are more likely to be the same (ECB updates on CET
afternoon).}

\item{use_most_recent}{\emph{if no data is available, should most recent data be used?}

\verb{scalar<logical>} // \emph{default:} \code{TRUE} (\code{optional})

\code{date} might be a Saturday, Sunday or the current day before EXR publication. If \code{use_most_recent}
is \code{TRUE}, the API will be asked for the most recent date before the given date}
}
\value{
a \code{tibble} with columns 'ISOCODE' and 'TITLE'
}
\description{
The available currencies at a specific date. For a full list of all currencies,
have a look at \href{https://www.iso.org/iso-4217-currency-codes.html}{ISO 4217}
or \link[ISOcodes:ISO_4217]{ISOcodes::ISO_4217}
}
\examples{

# get all available currencies for 2nd of Jan 2024
# (and don't look into historical data if there are no information available)
EXR::get_available_currencies(as.Date("2024-01-02"), use_most_recent = FALSE)

# get all currencies for the most recent date in the database
EXR::get_available_currencies(Sys.Date())

# get a vector of available ISO codes
as.vector(get_available_currencies()$ISOCODE)

}
