% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ecb_api_request.R
\name{perform_ecb_api_request}
\alias{perform_ecb_api_request}
\title{performs a GET-request against the ECB Data API}
\usage{
perform_ecb_api_request(query)
}
\arguments{
\item{query}{\emph{the EXR request with both key and parameters}

\verb{scalar<character>} // \strong{required}

A query compliant to the \href{https://data.ecb.europa.eu/help/api/data}{official documentation}.
A generic form of the whole API-request is
\itemize{
\item \code{protocol://wsEntryPoint/resource/flowRef/key?parameters}
}
but for this function, everything but \code{key?parameters} is already set automatically.

Example requests might be:
\itemize{
\item \code{A.USD.EUR.SP00.A?format=csvdata&startPeriod=2018&detail=dataonly}
\item \code{D.USD.EUR.SP00.A}
}}
}
\value{
\verb{scalar<character>}, the return value of \code{\link[httr2:resp_body_raw]{httr2::resp_body_string()}}. If body
is empty, NULL will be returned.
}
\description{
the functionality is constrained to EXR-data. It's primarily intended for
internal use. For access to the whole API with all ECB data but without any
further functionality have a look at the package \link[ecb:get_data]{ecb::get_data}.
}
\details{
The free API is not very fast, if you query large amounts of data, you might have
to increase the number. The actual amount of data is not limited but timeout is set to
\code{getOption("EXR.http.timeout", default = 30)} by this package. Therefore
you may set \code{options("EXR.http.timeout" = x)} to override the default value.
}
\examples{

# get annual average USD/EUR exchange rates in CSV-format as string, starting from 2018
EXR::perform_ecb_api_request("A.USD.EUR.SP00.A?format=csvdata&startPeriod=2018&detail=dataonly")

# get annual average USD/EUR exchange rates as table, starting from 2018, all columns
EXR::perform_ecb_api_request("A.USD.EUR.SP00.A?format=csvdata&startPeriod=2018") |>
  readr::read_csv(show_col_types = FALSE)

}
\references{
\href{https://data.ecb.europa.eu/help/api/data}{https://data.ecb.europa.eu/help/api/data}
}
