% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exr.R
\name{plot.exr}
\alias{plot.exr}
\title{Visualize the distribution of the CATEs and estimated external robustness}
\usage{
\method{plot}{exr}(x, cate_method = NULL, uncertainty = FALSE, ...)
}
\arguments{
\item{x}{the output from exr()}

\item{cate_method}{When "grf", we use 'grf'. When "X-learner", we use 'X-learner' based on R package 'SuperLearner'.}

\item{uncertainty}{Logical. Whether we visualize the distribution of the CATEs and estimated external robustness based on an adjusted threshold that takes into account uncertainties.}

\item{...}{Other arguments.}
}
\description{
Visualize the distribution of the CATEs and estimated external robustness
}
