% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sdmx_character_to_date}
\alias{sdmx_character_to_date}
\title{vectorised transformation from character (sdmx Reporting Periods) to date}
\usage{
sdmx_character_to_date(char, day_in_period = "last")
}
\arguments{
\item{char}{\emph{date strings}

\verb{scalar<character>} or \verb{vector<character>} // \strong{required}

Date representations such as "2012-Q2" for 2nd quarter of 2012.}

\item{day_in_period}{\emph{which day to return for a time period}

\verb{scalar<character>} //  \emph{default:} \code{last} (\code{optional})

which day to return for a time period. Can be either "first" or "last" for
the first day of a period or the last day. Calculation does not care about
working days.}
}
\value{
\verb{scalar<Date>} or \verb{vector<Date>}, depending on the input
}
\description{
The function determines the format automatically and returns the corresponding date.
In case of a time period such as quarters or years, you can choose to return the first
or last day of that period. sdmx notation (Statistical Data and Metadata eXchange)
}
\note{
intended for internal use
}
\examples{
# different values
EXR::sdmx_character_to_date(c("2024", "2024-08-20", "2023-S2", "2022-07", "2021-Q4", NA))

# different values, but return first day of period instead of last day
EXR::sdmx_character_to_date(
  char = c("2024", "2024-08-20", "2023-S2", "2022-07", "2021-Q4", NA),
  day_in_period = "first"
)

# one single transformation
EXR::sdmx_character_to_date("2023")
}
\references{
\itemize{
\item \href{https://data.ecb.europa.eu/help/api/data}{ECB API Documentation}
\item \href{https://sdmx.org/wp-content/uploads/sdmx_2-1_SECTION_6_TechnicalNotes.pdf}{sdmx Standards, technical notes}
}
}
\seealso{
\code{\link[=sdmx_date_to_character]{sdmx_date_to_character()}}
}
