% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sdmx_date_to_character}
\alias{sdmx_date_to_character}
\title{vectorised transformation from date to character (sdmx Reporting Periods)}
\usage{
sdmx_date_to_character(date, frequency)
}
\arguments{
\item{date}{\emph{date value}

\verb{scalar<Date>} or \verb{vector<Date>} // \strong{required}

Date values (\code{as.Date}). If length of vector is > 1 then it must be the same
size as param \code{frequency}. If length of vector is 1, then it will be scaled to
the size of frequency vector. sdmx notation (Statistical Data and Metadata eXchange)}

\item{frequency}{\emph{time related frequency of data}

\verb{scalar<character>} or \verb{vector<character>} // \strong{required}

allowed values are either A, D, H, M, Q
\itemize{
\item \strong{A}...nnualy
\item \strong{D}...aily
\item \strong{H}...alf-Yearly
\item \strong{M}...onthly
\item \strong{Q}...uarterly
}

If length of \code{frequency} is > 1 then it must be either the same size as
param \code{date} or \code{date} must be scalar. If length of vector is 1, then it will
be scaled to the size of \code{date} vector}
}
\value{
\verb{scalar<character>} or \verb{vector<character>}, depending on the input.
Vector is compliant to sdmx notation (Statistical Data and Metadata eXchange)
}
\description{
vectorised transformation from date to character (sdmx Reporting Periods)
}
\note{
intended for internal use
}
\examples{
# corresponding vectors
EXR::sdmx_date_to_character(
  date = as.Date(c("2024-12-31", "2024-08-20", "2023-12-31", "2022-07-31", "2021-12-31", NA)),
  frequency = c("A", "D", "H", "M", "Q", "D")
)

# today's quarter
EXR::sdmx_date_to_character(Sys.Date(), "Q")

# multiple formats for today
EXR::sdmx_date_to_character(Sys.Date(), c("A", "D", "H", "M", "Q"))

# date vector with static conversion
EXR::sdmx_date_to_character(
  date = as.Date(c("2024-12-31", "2024-08-20", "2023-12-31", "2022-07-31", "2021-12-31", NA)),
  frequency = "A"
)

}
\references{
\itemize{
\item \href{https://data.ecb.europa.eu/help/api/data}{ECB API Documentation}
\item \href{https://sdmx.org/wp-content/uploads/sdmx_2-1_SECTION_6_TechnicalNotes.pdf}{sdmx Standards, technical notes}
}
}
\seealso{
\code{\link[=sdmx_character_to_date]{sdmx_character_to_date()}}
}
\keyword{internal}
