% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_exr.R
\name{summary.exr}
\alias{summary.exr}
\title{Summary function}
\usage{
\method{summary}{exr}(object, digits = 2, ...)
}
\arguments{
\item{object}{the output from exr()}

\item{digits}{digits used for a table of covariate profiles. Default = 2.}

\item{...}{Other arguments.}
}
\description{
Summary function
}
