% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable_exr.R
\name{xtable.exr}
\alias{xtable.exr}
\title{Function to produce a LaTeX output}
\usage{
\method{xtable}{exr}(obj, digits = 2, ...)
}
\arguments{
\item{digits}{digits used for a table of covariate profiles. Default = 2.}

\item{...}{Other arguments.}

\item{object}{the output from exr()}
}
\description{
Function to produce a LaTeX output
}
